Added Tests. Fixed train/fullpop
Added a new testing file to visualize errors better.

Also renamed all input/output nodes in the training full pop so I stop getting confused.

+++ pycharm_project/network.py
@@ -102,7 +102,7 @@ class Network:
 
         error = target - outputs
 
-        output_error_gradients = self.__convert_to_gradients(error)
+        output_error_gradients = self.__convert_to_gradients(error.astype(np.float32))
 
         gradient_data = [[None, None] for j in range(len(self.layout))]
